update profile-image styles for safari
apparently the inset and outset styles don't look the same on macos and ios safari. They seem broken and I'm not sure how to fix this at the moment without creating an svg shape or an image. So for now I'm just reverting to a simpler style of solid borders.